requirements.txt
altair==5.5.0
attrs==24.3.0
azure-core==1.32.0
azure-identity==1.19.0
azure-keyvault-secrets==4.9.0
azure-storage-blob==12.24.0
babel==2.16.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
dateparser==1.2.0
docx==0.2.4
docxcompose==1.4.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
isodate==0.7.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msal==1.31.1
msal-extensions==1.2.0
narwhals==1.20.1
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pathvalidate==3.2.2
pillow==11.1.0
portalocker==2.10.1
protobuf==5.29.2
pyarrow==18.1.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
setuptools==75.6.0
six==1.17.0
smmap==5.0.2
streamlit==1.41.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
validators==0.34.0
